+++ gateway-js/src/datasources/types.ts
@@ -1,3 +1,4 @@
+import { ResponsePath } from '@apollo/query-planner';
 import { GatewayGraphQLResponse, GatewayGraphQLRequestContext } from '@apollo/server-gateway-interface';
 
 export interface GraphQLDataSource<
@@ -47,6 +48,12 @@ export type GraphQLDataSourceProcessOptions<
        * this will increase the memory used by the gateway query plan cache.
        */
       document?: GatewayGraphQLRequestContext<TContext>['document'];
+
+      /**
+      * The path in the overall gateway operation at which that subgraph request gets inserted.
+      * Please note that this could be set to `undefined` when the path is not available, or set to an empty array for top-level fetch operations.
+      */
+      pathInIncomingRequest?: ResponsePath;
     }
   | {
       kind: