+++ query-graphs-js/src/graphPath.ts
@@ -206,7 +206,7 @@ export class GraphPath<TTrigger, RV extends Vertex = Vertex, TNullEdge extends n
       edgeConditions: [],
       ownPathIds: [],
       overriddingPathIds: [],
-      runtimeTypesOfTail: runtimeTypes
+      runtimeTypesOfTail: runtimeTypes,
     });
   }
 
@@ -278,8 +278,21 @@ export class GraphPath<TTrigger, RV extends Vertex = Vertex, TNullEdge extends n
     return this.props.edgeTriggers[this.size - 1];
   }
 
+  /** The possible runtime types the tail of the path can be (this is deduplicated). */
   tailPossibleRuntimeTypes(): readonly ObjectType[] {
     return this.props.runtimeTypesOfTail;
+  }
+
+  /**
+   * Returns `true` if the last edge of the path correspond to an @interfaceObject "fake cast" while the the previous edge was an edge that "entered" the subgraph (a key edge from another subgraph).
+   */
+  lastIsIntefaceObjectFakeDownCastAfterEnteringSubgraph(): boolean {
+    return this.lastIsInterfaceObjectFakeDownCast()
+      && this.subgraphEnteringEdge?.index === this.size - 2; // size - 1 is the last index (the fake cast), so size - 2 is the previous edge.
+  }
+
+  private lastIsInterfaceObjectFakeDownCast(): boolean {
+    return this.lastEdge()?.transition.kind === 'InterfaceObjectFakeDownCast';
   }
 
   /**
@@ -340,7 +353,9 @@ export class GraphPath<TTrigger, RV extends Vertex = Vertex, TNullEdge extends n
                 edgeConditions: withReplacedLastElement(this.props.edgeConditions, conditionsResolution.pathTree ?? null),
                 edgeToTail: updatedEdge,
                 runtimeTypesOfTail: runtimeTypesWithoutPreviousCast,
-                deferOnTail: defer,
+                // We know the edge is a DownCast, so if there is no new `defer` taking precedence, we just inherit the
+                // prior version.
+                deferOnTail: defer ?? this.props.deferOnTail,
               });
             }
           }
@@ -365,13 +380,14 @@ export class GraphPath<TTrigger, RV extends Vertex = Vertex, TNullEdge extends n
         // type is not an @interfaceObject itself, then we can eliminate that last edge as it does nothing useful, but also,
         // it has conditions and we don't need/want the key we're following to depend on those conditions, since it doesn't have
         // to.
-        if (this.props.edgeToTail?.transition.kind === 'InterfaceObjectFakeDownCast' && isInterfaceType(edge.tail.type)) {
+        if (this.lastIsInterfaceObjectFakeDownCast() && isInterfaceType(edge.tail.type)) {
           return new GraphPath({
             ...this.props,
             tail: edge.tail,
             edgeTriggers: withReplacedLastElement(this.props.edgeTriggers, trigger),
             edgeIndexes: withReplacedLastElement(this.props.edgeIndexes, edge.index),
             edgeConditions: withReplacedLastElement(this.props.edgeConditions, conditionsResolution.pathTree ?? null),
+            subgraphEnteringEdge,
             edgeToTail: edge,
             runtimeTypesOfTail: updateRuntimeTypes(this.props.runtimeTypesOfTail, edge),
             runtimeTypesBeforeTailIfLastIsCast: undefined, // we know last is not a cast
@@ -391,7 +407,10 @@ export class GraphPath<TTrigger, RV extends Vertex = Vertex, TNullEdge extends n
       edgeToTail: edge,
       runtimeTypesOfTail: updateRuntimeTypes(this.props.runtimeTypesOfTail, edge),
       runtimeTypesBeforeTailIfLastIsCast: edge?.transition?.kind === 'DownCast' ? this.props.runtimeTypesOfTail : undefined,
-      deferOnTail: defer,
+      // If there is no new `defer` taking precedence, and the edge is downcast, then we inherit the prior version. This
+      // is because we only try to re-enter subgraphs for @defer on concrete fields, and so as long as we add downcasts,
+      // we should remember that we still need to try re-entering the subgraph.
+      deferOnTail: defer ?? (edge && edge.transition.kind === 'DownCast' ? this.props.deferOnTail : undefined),
     });
   }
 
@@ -430,7 +449,7 @@ export class GraphPath<TTrigger, RV extends Vertex = Vertex, TNullEdge extends n
     });
   }
 
-  checkDirectPathFomPreviousSubgraphTo(
+  checkDirectPathFromPreviousSubgraphTo(
     typeName: string,
     triggerToEdge: (graph: QueryGraph, vertex: Vertex, t: TTrigger) => Edge | null | undefined
   ): Vertex | undefined {
@@ -851,7 +870,7 @@ export class TransitionPathWithLazyIndirectPaths<V extends Vertex = Vertex> {
 // the actual tree, which we need for query planning, or simply returning "undefined" which means "The condition can be satisfied but I didn't
 // bother computing a tree for it", which we use for simple validation.
 
-// Returns undefined if there is no way to advance the path with this transition. Otherwise, it returns a list of options (paths) we can be in after advancing the transition.
+// Returns some a `Unadvanceables` object if there is no way to advance the path with this transition. Otherwise, it returns a list of options (paths) we can be in after advancing the transition.
 // The lists of options can be empty, which has the special meaning that the transition is guaranteed to have no results (it corresponds to unsatisfiable conditions),
 // meaning that as far as composition validation goes, we can ignore that transition (and anything that follows) and otherwise continue.
 export function advancePathWithTransition<V extends Vertex>(
@@ -1110,6 +1129,25 @@ function advancePathWithNonCollectingAndTypePreservingTransitions<TTrigger, V ex
   convertTransitionWithCondition: (transition: Transition, context: PathContext) => TTrigger,
   triggerToEdge: (graph: QueryGraph, vertex: Vertex, t: TTrigger) => Edge | null | undefined
 ): IndirectPaths<TTrigger, V, TNullEdge, TDeadEnds>  {
+  // If we're asked for indirect paths after an "@interfaceObject fake down cast" but that down cast comes just after a non-collecting edges, then
+  // we can ignore it (skip indirect paths from there). The reason is that the presence of the non-collecting just before the fake down-cast means
+  // we add looked at indirect paths just before that down cast, but that fake downcast really does nothing in practice with the subgraph it's on,
+  // so any indirect path from that fake down cast will have a valid indirect path _before_ it, and so will have been taken into account independently.
+  if (path.lastIsIntefaceObjectFakeDownCastAfterEnteringSubgraph()) {
+    // Note: we need to register a dead-end for every subgraphs we "could" be going to, or the code calling this may try to infer a reason on its own
+    // and we'll run into some assertion.
+    const reachableSubgraphs = new Set(path.nextEdges().filter((e) => !e.transition.collectOperationElements && e.tail.source !== path.tail.source).map((e) => e.tail.source));
+    return {
+      paths: [],
+      deadEnds: new Unadvanceables(Array.from(reachableSubgraphs).map((s) => ({
+        sourceSubgraph: path.tail.source,
+        destSubgraph: s,
+        reason: UnadvanceableReason.IGNORED_INDIRECT_PATH,
+        details: `ignoring moving from "${path.tail.source}" to "${s}" as a more direct option exists`,
+      }))) as TDeadEnds,
+    };
+  }
+
   const isTopLevelPath = path.isOnTopLevelQueryRoot();
   const typeName = isFederatedGraphRootType(path.tail.type) ? undefined : path.tail.type.name;
   const originalSource = path.tail.source;
@@ -1245,7 +1283,7 @@ function advancePathWithNonCollectingAndTypePreservingTransitions<TTrigger, V ex
         // loop when calling `hasValidDirectKeyEdge` in that case without additional care and it's not useful because this
         // very method already ensure we don't create unnecessary chains of keys for the "current type"
         if (subgraphEnteringEdge && edge.transition.kind === 'KeyResolution' && subgraphEnteringEdge.edge.tail.type.name !== typeName) {
-          const prevSubgraphVertex = toAdvance.checkDirectPathFomPreviousSubgraphTo(edge.tail.type.name, triggerToEdge);
+          const prevSubgraphVertex = toAdvance.checkDirectPathFromPreviousSubgraphTo(edge.tail.type.name, triggerToEdge);
           const backToPreviousSubgraph = subgraphEnteringEdge.edge.head.source === edge.tail.source;
           const maxCost = toAdvance.subgraphEnteringEdge.cost + (backToPreviousSubgraph ? 0 : conditionResolution.cost);
           if (prevSubgraphVertex
@@ -1711,9 +1749,10 @@ export function advanceSimultaneousPathsWithOperation<V extends Vertex>(
     let options: SimultaneousPaths<V>[] | undefined = undefined;
 
     debug.group(() => `Computing options for ${path}`);
+    const shouldReenterSubgraph = path.deferOnTail && operation.kind === 'Field';
     // If we're just entering a deferred section, then we will need to re-enter subgraphs, so we should not consider
     // direct options and instead force an indirect path.
-    if (!path.deferOnTail) {
+    if (!shouldReenterSubgraph) {
       debug.group(() => `Direct options`);
       const { options: advanceOptions, hasOnlyTypeExplodedResults } = advanceWithOperation(
         supergraphSchema,
@@ -1749,41 +1788,43 @@ export function advanceSimultaneousPathsWithOperation<V extends Vertex>(
 
     // If there was not valid direct path (or we didn't check those because we enter a defer), that's ok, we'll just try with non-collecting edges.
     options = options ?? [];
-    debug.group(`Computing indirect paths:`);
-    // Then adds whatever options can be obtained by taking some non-collecting edges first.
-    const pathsWithNonCollecting = subgraphSimultaneousPaths.indirectOptions(updatedContext, i);
-    debug.groupEnd(() => pathsWithNonCollecting.paths.length == 0 ? `no indirect paths` : `${pathsWithNonCollecting.paths.length} indirect paths`);
-    if (pathsWithNonCollecting.paths.length > 0) {
-      debug.group('Validating indirect options:');
-      for (const pathWithNonCollecting of pathsWithNonCollecting.paths) {
-        debug.group(() => `For indirect path ${pathWithNonCollecting}:`);
-        const { options: pathWithOperation } = advanceWithOperation(
-          supergraphSchema,
-          pathWithNonCollecting,
-          operation,
-          updatedContext,
-          subgraphSimultaneousPaths.conditionResolver
-        );
-        // If we can't advance the operation after that path, ignore it, it's just not an option.
-        if (!pathWithOperation) {
-          debug.groupEnd(() => `Ignoring: cannot be advanced with ${operation}`);
-          continue;
+    if (operation.kind === 'Field') {
+      debug.group(`Computing indirect paths:`);
+      // Then adds whatever options can be obtained by taking some non-collecting edges first.
+      const pathsWithNonCollecting = subgraphSimultaneousPaths.indirectOptions(updatedContext, i);
+      debug.groupEnd(() => pathsWithNonCollecting.paths.length == 0 ? `no indirect paths` : `${pathsWithNonCollecting.paths.length} indirect paths`);
+      if (pathsWithNonCollecting.paths.length > 0) {
+        debug.group('Validating indirect options:');
+        for (const pathWithNonCollecting of pathsWithNonCollecting.paths) {
+          debug.group(() => `For indirect path ${pathWithNonCollecting}:`);
+          const { options: pathWithOperation } = advanceWithOperation(
+            supergraphSchema,
+            pathWithNonCollecting,
+            operation,
+            updatedContext,
+            subgraphSimultaneousPaths.conditionResolver
+          );
+          // If we can't advance the operation after that path, ignore it, it's just not an option.
+          if (!pathWithOperation) {
+            debug.groupEnd(() => `Ignoring: cannot be advanced with ${operation}`);
+            continue;
+          }
+          debug.groupEnd(() => `Adding valid option: ${pathWithOperation}`);
+          // advancedWithOperation can return an empty list only if the operation if a fragment with a condition that, on top of the "current" type
+          // is unsatisfiable. But as we've only taken type preserving transitions, we cannot get an empty results at this point if we haven't
+          // had one when testing direct transitions above (in which case we have exited the method early).
+          assert(pathWithOperation.length > 0, () => `Unexpected empty options after non-collecting path ${pathWithNonCollecting} for ${operation}`);
+          options = options.concat(pathWithOperation);
         }
-        debug.groupEnd(() => `Adding valid option: ${pathWithOperation}`);
-        // advancedWithOperation can return an empty list only if the operation if a fragment with a condition that, on top of the "current" type
-        // is unsatisfiable. But as we've only taken type preserving transitions, we cannot get an empty results at this point if we haven't
-        // had one when testing direct transitions above (in which case we have exited the method early).
-        assert(pathWithOperation.length > 0, () => `Unexpected empty options after non-collecting path ${pathWithNonCollecting} for ${operation}`);
-        options = options.concat(pathWithOperation);
+        debug.groupEnd();
       }
-      debug.groupEnd();
     }
 
     // If we were entering a @defer, we've skipped the potential "direct" options because we need an "indirect" one (a key/root query)
     // to be able to actualy defer. But in rare cases, it's possible we actually couldn't resolve the key fields needed to take a key
     // but could still find a direct path. If so, it means it's a corner case where we cannot do query-planner-based-@defer and have
     // to fall back on not deferring.
-    if (options.length === 0 && path.deferOnTail) {
+    if (options.length === 0 && shouldReenterSubgraph) {
       debug.group(() => `Cannot defer (no indirect options); falling back to direct options`);
       const { options: advanceOptions } = advanceWithOperation(
         supergraphSchema,
@@ -2089,6 +2130,12 @@ function advanceWithOperation<V extends Vertex>(
         );
         return { options: pathAsOptions(fieldPath) };
       case 'InterfaceType':
+        // Due to @interfaceObject, we could be in a case where the field asked is not on the interface but
+        // rather on one of it's implementation. This can happen if we just entered the subgraph on an interface @key
+        // and coming from an @interfaceObject. In that case, we'll skip checking for an interface direct edge and
+        // simply cast into that implementation below.
+        const fieldIsOfAnImplementation = field.parent.name !== currentType.name;
+
         // First, we check if there is a direct edge from the interface (which only happens if we're in a subgraph that knows all of the
         // implementations of that interface globally and all of them resolve the field).
         // If there is one, then we have 2 options:
@@ -2100,7 +2147,7 @@ function advanceWithOperation<V extends Vertex>(
         // - either type-exploding cannot work unless taking the interface edge also do (the `anImplementationIsEntityWithFieldShareable`)
         // - or that type-exploding cannot be more efficient than the direct path (when no @provides are involved; if a provide is involved
         //   in one of the implementation, then type-exploding may lead to a shorter overall plan thanks to that @provides)
-        const itfEdge = nextEdgeForField(path, operation);
+        const itfEdge = fieldIsOfAnImplementation ? undefined : nextEdgeForField(path, operation);
         let itfPath: OpGraphPath<V> | undefined = undefined;
         let directPathOverrideTypeExplosion = false;
         if (itfEdge) {
@@ -2133,12 +2180,11 @@ function advanceWithOperation<V extends Vertex>(
         // - the most common is that it's a field of the interface that is queried, and
         //   so we should type-explode because either didn't had a direct edge, or @provides
         //   makes it potentially worthwile to check with type explosion.
-        // - but we could also have the case where the field queried is actually of one
-        //   of the implementation of the interface: this happens if we just entered the
-        //   subgraph on an interface @key. In that case, we only want to consider that one
+        // - but, as mentionned earlier, we could be in the case where the field queried is actually of one
+        //   of the implementation of the interface. In that case, we only want to consider that one
         //   implementation.
         let implementations: readonly ObjectType[];
-        if (field.parent.name !== currentType.name) {
+        if (fieldIsOfAnImplementation) {
           assert(
             isObjectType(field.parent) && path.tailPossibleRuntimeTypes().some((t) => t.name === field.parent.name),
             () => `${field.coordinate} requested on ${currentType}, but ${field.parent} is not an implementation`